requirements.txt
amqp==2.6.1
asgiref==3.2.10
asyncio==3.4.3
beautifulsoup4==4.9.1
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.0.8
django-import-export==2.3.0
django-jsonfield==1.4.0
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
jdcal==1.4.1
kombu==4.6.11
lxml==4.5.2
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.5
psycopg2-binary==2.8.5
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tablib==2.0.0
urllib3==1.25.9
vine==1.3.0
whitenoise==5.1.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
